User Stories D2 R1 backend + E1, E2, E4 frontend bugfix
- FIXED major bug with items, now price is of type double and id is of type long
- FIXED now cashier view shows pop-up errors given an exception
- FIXED typo and small code irregularities

+++ Client/src/mediator/NetworkManager.java
@@ -141,7 +141,7 @@ public class NetworkManager implements RemoteModel {
     }
 
     @Override
-    public void changePrice(int id, int price) throws RemoteException {
+    public void changePrice(Long id, Double price) throws RemoteException {
         remoteModel.changePrice(id, price);
     }
