+++ Client/src/model/InventoryModel.java
@@ -4,5 +4,6 @@ import java.util.ArrayList;
 
 public interface InventoryModel {
     ArrayList<Item> getItems();
-    void changePrice(int id, int price);
+
+    void changePrice(Long id, Double price);
 }
